Promote SegmentStatsTracker to top-level class, add tests
